backend/requirements.txt
aniso8601==6.0.0
astroid==2.4.2
attrs==19.3.0
Click==7.0
Flask==1.0.3
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
pylint==2.5.3
pyparsing==2.4.7
pytest==5.4.3
python-dotenv==0.14.0
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.4
toml==0.10.1
wcwidth==0.2.5
Werkzeug==0.15.5
wrapt==1.12.1
